+++ example/cli/src/app/app.module.ts
@@ -11,12 +11,17 @@ import { FullTreeComponent } from './fulltree/fulltree.component';
 import { TemplatesComponent } from './templates/templates.component';
 import { FilterComponent } from './filter/filter.component';
 import { FieldsComponent } from './fields/fields.component';
+import { RtlTreeComponent } from './rtl/rtl-tree.component';
+import { AsyncTreeComponent } from './async/async.component';
+import { SaveRestoreComponent } from './save-restore/save-restore.component';
+import { CheckboxesComponent } from './checkboxes/checkboxes.component';
 
 /*
   Turn on to check if tree supports strict mode in MobX.
   But remember to turn off (to allow users of MobX not to use strict mode in their apps)
 */
 import { useStrict } from 'mobx';
+import { EmptyComponent } from './empty/empty.component';
 useStrict(true);
 
 
@@ -27,7 +32,12 @@ useStrict(true);
     FullTreeComponent,
     TemplatesComponent,
     FieldsComponent,
-    FilterComponent
+    FilterComponent,
+    EmptyComponent,
+    RtlTreeComponent,
+    AsyncTreeComponent,
+    SaveRestoreComponent,
+    CheckboxesComponent
   ],
   imports: [
     BrowserModule,