+++ lib/angular-tree-component.ts
@@ -4,7 +4,7 @@ import { MobxAngularModule } from 'mobx-angular';
 import { MaterialModule } from '@angular/material';
 
 import { TREE_ACTIONS, IActionMapping, IActionHandler } from './models/tree-options.model';
-import { ITreeOptions, IAllowDropFn, IAllowDragFn } from './defs/api';
+import { ITreeOptions, IAllowDropFn, IAllowDragFn, ITreeState } from './defs/api';
 import { KEYS } from './constants/keys';
 import { TreeModel } from './models/tree.model';
 import { TreeNode } from './models/tree-node.model';
@@ -48,7 +48,8 @@ export {
   TreeNodeChildrenComponent,
   TreeNodeDropSlot,
   TreeNodeCollectionComponent,
-  TreeViewportComponent
+  TreeViewportComponent,
+  ITreeState
 };
 
 @NgModule({